+++ FE/src/components/common/header/Modal/UserActiveModal.loader.tsx
@@ -1,4 +1,4 @@
-import LoadingSpinner from "../LoadingSpinner";
+import LoadingSpinner from "../../LoadingSpinner";
 
 const UserActiveModalSkeleton = () => {
   return (
